requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.3
packaging==24.1
pillow==11.0.0
propcache==0.2.0
pyparsing==3.2.0
pyrtlsdr==0.3.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
